+++ cohttp-async/src/client.mli
@@ -1,18 +1,26 @@
 (** Send an HTTP request with an arbitrary body
     The request is sent as-is. *)
 val request :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?uri:Uri.t ->
   ?body:Body.t ->
   Cohttp.Request.t ->
   (Cohttp.Response.t * Body.t) Async_kernel.Deferred.t
 
 (** Send an HTTP request with arbitrary method and a body
-    Infers the transfer encoding *)
+    Infers the transfer encoding. Depending on the given [uri],
+    we choose a way to start a communication such as:
+
+    {ul
+    {- If the scheme is [https], we try to initiate an SSL connection with
+       the given [ssl_ctx] or a default one on the default port ([*:443]) or
+       the specified one.}
+    {- If the scheme is [httpunix], we use a UNIX domain socket.}
+    {- If the scheme ie [http], we try an usual TCP/IP connection on the
+       default port ([*:80]) or the specified one.}}
+*)
 val call :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   ?chunked:bool ->
   ?body:Body.t ->
@@ -20,33 +28,49 @@ val call :
   Uri.t ->
   (Cohttp.Response.t * Body.t) Async_kernel.Deferred.t
 
+
+module Connection : sig
+  type t
+
+  val connect :
+    ?ssl_ctx:Conduit_async_ssl.context ->
+    Uri.t ->
+    t Async_kernel.Deferred.t
+
+  val close : t -> unit Async_kernel.Deferred.t
+
+  val is_closed : t -> bool
+
+  val request :
+    ?body: Body.t ->
+    t ->
+    Cohttp.Request.t ->
+    (Cohttp.Response.t * Body.t) Async_kernel.Deferred.t
+end
+
 val callv :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   Uri.t ->
   (Cohttp.Request.t * Body.t) Async_kernel.Pipe.Reader.t ->
   (Cohttp.Response.t * Body.t) Async_kernel.Pipe.Reader.t Async_kernel.Deferred.t
 
 (** Send an HTTP GET request *)
 val get :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   Uri.t ->
   (Cohttp.Response.t * Body.t) Async_kernel.Deferred.t
 
 (** Send an HTTP HEAD request *)
 val head :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   Uri.t ->
   Cohttp.Response.t Async_kernel.Deferred.t
 
 (** Send an HTTP DELETE request *)
 val delete :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   ?chunked:bool ->
   ?body:Body.t ->
@@ -57,8 +81,7 @@ val delete :
     [chunked] encoding is off by default as not many servers support it
 *)
 val post :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   ?chunked:bool ->
   ?body:Body.t ->
@@ -69,8 +92,7 @@ val post :
     [chunked] encoding is off by default as not many servers support it
 *)
 val put :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   ?chunked:bool ->
   ?body:Body.t ->
@@ -81,8 +103,7 @@ val put :
     [chunked] encoding is off by default as not many servers support it
 *)
 val patch :
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   ?chunked:bool ->
   ?body:Body.t ->
@@ -91,8 +112,7 @@ val patch :
 
 (** Send an HTTP POST request in form format *)
 val post_form:
-  ?interrupt:unit Async_kernel.Deferred.t ->
-  ?ssl_config:Conduit_async.V2.Ssl.Config.t ->
+  ?ssl_ctx:Conduit_async_ssl.context ->
   ?headers:Cohttp.Header.t ->
   params:(string * string list) list ->
   Uri.t ->