+++ cohttp-lwt-unix/src/net.mli
@@ -16,21 +16,28 @@
 
 (** Basic satisfaction of {! Cohttp_lwt.Net } *)
 
-module IO = Io
+module IO = Cohttp_lwt_unix_nossl.IO
 
-type ctx = {
-  ctx : Conduit_lwt_unix.ctx;
-  resolver : Resolver_lwt.t;
-} [@@deriving sexp_of]
-
-val init : ?ctx:Conduit_lwt_unix.ctx -> ?resolver:Resolver_lwt.t -> unit -> ctx
+type ctx = (Conduit.resolvers[@sexp.opaque]) [@@deriving sexp]
 
 val default_ctx : ctx
 
 val connect_uri :
   ctx:ctx ->
   Uri.t ->
-  (Conduit_lwt_unix.flow * Conduit_lwt_unix.ic * Conduit_lwt_unix.oc) Lwt.t
+  (Conduit_lwt.flow * Lwt_io.input Lwt_io.channel * Lwt_io.output Lwt_io.channel) Lwt.t
+(** [connect_uri ~ctx uri] starts a {i flow} on the given [uri]. The choice of the
+   protocol (with or without encryption) is done by the {i scheme} of the given [uri]:
+
+    {ul
+    {- If the scheme is [https], we will {b extend} [ctx] to be able to start a TLS connection
+       with a default TLS configuration (no authentication) on the default or user-specified port.}
+    {- If the scheme is [http], we will {b extend} [ctx] to be able to start a simple TCP/IP
+       connection on the default or user-specified port.}}
+
+    These extensions have the highest priority ([Conduit] will try to initiate a communication with
+   them first). By {i extension}, we mean that the user is able to fill its own [ctx] and we don't
+   overlap resolution functions from the given [ctx]. *)
 
 val close_in : 'a Lwt_io.channel -> unit
 val close_out : 'a Lwt_io.channel -> unit