+++ cohttp-lwt/src/s.ml
@@ -21,8 +21,10 @@ end
     and close the resulting channels to clean up. *)
 module type Net = sig
   module IO : IO
-  type ctx [@@deriving sexp_of]
-  val default_ctx: ctx
+
+  type ctx [@@deriving sexp]
+
+  val default_ctx : ctx
   val connect_uri : ctx:ctx -> Uri.t -> (IO.conn * IO.ic * IO.oc) Lwt.t
   val close_in : IO.ic -> unit
   val close_out : IO.oc -> unit
@@ -36,13 +38,11 @@ end
     fashion.  It will still be finalized by a GC hook if it is not used
     up, but this can take some additional time to happen. *)
 module type Client = sig
-
-  type ctx [@@deriving sexp_of]
-  val default_ctx : ctx
+  type ctx
 
   (** [call ?ctx ?headers ?body ?chunked meth uri] will resolve the
-      [uri] to a concrete network endpoint using the resolver initialized
-      in [ctx].  It will then issue an HTTP request with method [meth],
+      [uri] to a concrete network endpoint using the {!Conduit.resolvers} [ctx].
+      It will then issue an HTTP request with method [meth],
       adding request headers from [headers] if present.  If a [body]
       is specified then that will be included with the request, using
       chunked encoding if [chunked] is true.  The default is to disable
@@ -50,7 +50,14 @@ module type Client = sig
 
       In most cases you should use the more specific helper calls in the
       interface rather than invoke this function directly.  See {!head},
-      {!get} and {!post} for some examples. *)
+      {!get} and {!post} for some examples.
+  
+      Depending on [ctx], the library is able to send a simple HTTP request
+      or an encrypted one with a secured protocol (such as TLS). By default
+      (on [cohttp-lwt-unix]), [ctx] tries to initiate a secured connection
+      with TLS (it uses [ocaml-tls]) on [*:443] or on the specified port by
+      the user. If the peer is not available, [cohttp]/[conduit] tries the usual
+      ([*:80]) or the specified port by the user in a non-secured way. *)
   val call :
     ?ctx:ctx ->
     ?headers:Cohttp.Header.t ->