+++ cohttp/src/header.mli
@@ -19,46 +19,61 @@
     of values associated with a single key. *)
 type t [@@deriving sexp]
 
-(** Construct a fresh, empty map of HTTP headers *)
+(** Construct a fresh, empty map of HTTP headers. *)
 val init : unit -> t
 
-(** Test whether a HTTP headers are empty or not. *)
+(** Test whether HTTP headers are empty or not. *)
 val is_empty : t -> bool
 
-(** Construct a fresh map of HTTP headers with a single key and value entry *)
+(** Construct a fresh map of HTTP headers with a single key and value entry. *)
 val init_with  : string -> string -> t
 
-(** Add a key and value to an existing header map *)
+(** Add a key and value to an existing header map. *)
 val add : t -> string -> string -> t
 
-(** Add multiple key and value pairs to an existing header map *)
+(** Add multiple key and value pairs to an existing header map. *)
 val add_list : t -> (string * string) list -> t
 
-(** Add multiple values to a key in an existing header map *)
+(** Add multiple values to a key in an existing header map. *)
 val add_multi : t -> string -> string list -> t
 
 (** Given an optional header, either update the existing one with
     a key and value, or construct a fresh header with those values if
-    the header is [None] *)
+    the header is [None]. *)
 val add_opt : t option -> string -> string -> t
 
 (** Given a header, update it with the key and value unless the key is
-    already present in the header *)
+    already present in the header. *)
 val add_unless_exists : t -> string -> string -> t
 
-(** [add_unless_exists h k v] updates [h] with the key [k] and value [v]
+(** [add_opt_unless_exists h k v] updates [h] with the key [k] and value [v]
     unless the key is already present in the header.  If [h] is [None]
     then a fresh header is allocated containing the key [k] and the
     value [v]. *)
 val add_opt_unless_exists : t option -> string -> string -> t
 
-(** Remove a key from the header map and return a fresh header set.  The
+(** Remove a key from the header map and return a fresh header set. The
     original header parameter is not modified. *)
 val remove : t -> string -> t
 
-(** Replace a key from the header map if it exists.  The original
-    header parameter is not modified. *)
+(** Replace the value of a key from the header map if it exists, otherwise it
+    adds it to the header map. The original header parameter is not modified. *)
 val replace : t -> string -> string -> t
+
+(** [update h k f] returns a map containing the same headers as [h],
+    except for the header [k]. Depending on the value of [v] where [v] is
+    [f (get h k)], the header [k] is added, removed or updated.
+    If [v] is [None], the header is removed if it exists; otherwise,
+    if [v] is [Some z] then [k] is associated to [z] in the resulting headers.
+    If [k] was already associated in [h] to a value that is physically equal
+    to [z], [h] is returned unchanged. Similarly as for [get], if the header is
+    one of the set of headers defined to have list values, then all of the values
+    are concatenated into a single string separated by commas and passed to [f],
+    while the return value of [f] is split on commas and associated to [k].
+    If it is a singleton header, then the first value is passed to [f] and
+    no concatenation is performed, similarly for the return value.
+    The original header parameters are not modified. *)
+val update: t -> string -> (string option -> string option) -> t
 
 (** Check if a key exists in the header. *)
 val mem : t -> string -> bool