+++ cohttp/test/test_header.ml
@@ -133,6 +133,50 @@ let many_headers () =
   let h = add_header size (H.init ()) in
   Alcotest.(check int) "many_headers" (List.length (H.to_list h)) size
 
+module Updates = struct
+  let h = H.init ()
+    |> fun h -> H.add h "first" "1"
+    |> fun h -> H.add h "second" "2"
+    |> fun h -> H.add h "accept" "foo"
+    |> fun h -> H.add h "accept" "bar"
+
+  let replace_headers_if_exists () =
+    let h = H.replace h "second" "2a" in
+    Alcotest.(check (option string)) "replace_existing_header" (Some "2a") (H.get h "second")
+
+  let replace_headers_if_absent () =
+    let h = H.replace h "third" "3" in
+    Alcotest.(check (option string)) "replace_new_header" (Some "3") (H.get h "third")
+
+  let update_headers_if_exists () =
+    let h1 = H.update h "second" (function | Some _ -> Some "2a" | None -> None) in
+    let h2 = H.replace h "second" "2a" in
+    Alcotest.(check t_header) "update_existing_header" h1 h2
+  
+  let update_headers_if_exists_rm () =
+    let h1 = H.update h "second" (function | Some _ -> None | None -> Some "3") in
+    let h2 = H.remove h "second" in
+    Alcotest.(check t_header) "update_remove_header" h1 h2
+  
+  let update_headers_if_absent_add () =
+    let h = H.update h "third" (function | Some _ -> None | None -> Some "3") in
+    Alcotest.(check (option string)) "update_add_new_header" (Some "3") (H.get h "third")
+  
+  let update_headers_if_absent_rm () =
+    let h1 = H.update h "third" (function _ -> None) in
+    Alcotest.(check t_header) "update_remove_absent_header" h h1
+
+  let update_headers_if_exists_multi () =
+    let h1 = H.update h "accept" (function | Some v -> Some ("baz,"^v) | None -> None) in
+    let h2 = H.add h "accept" "baz" in
+    Alcotest.(check (option string)) "update_existing_header_multivalued" (H.get h1 "accept") (H.get h2 "accept")
+
+  let update_headers_if_absent () =
+    let h1 = H.update h "third" (function | Some _ -> Some "3" | None -> None) in
+    Alcotest.(check t_header) "update_new_header: unchanged" h h1;
+    Alcotest.(check (option string)) "update_new_header: map unchanged" None (H.get h "third")
+end
+
 module Content_range = struct
   let h1 = H.of_list ["Content-Length", "123"]
   let h2 = H.of_list ["Content-Range", "bytes 200-300/1000"]
@@ -441,50 +485,69 @@ let test_cachecontrol_concat () =
   aeso "test_cachecontrol_concat"
     (Some "public,max-age:86400") (H.get h "Cache-Control")
 
-;;
-Printexc.record_backtrace true;
-Alcotest.run "test_header" [
-  "Link", [
-    "simple", `Quick, link_simple;
-    "multiple rels", `Quick, link_multi_rel;
-    "multiple lines", `Quick, link_multi_line;
-    "multiheader", `Quick, link_multi_multi;
-    "rel uri", `Quick, link_rel_uri;
-    "anchor", `Quick, link_anchor;
-    "rev", `Quick, link_rev;
-    "media", `Quick, link_media;
-    "media complex", `Quick, link_media_complex;
-    "title", `Quick, link_title;
-    "title star", `Quick, link_title_star;
-    "type token", `Quick, link_type_token;
-    "type quoted", `Quick, link_type_quoted;
-    "extension", `Quick, link_ext;
-    "extension star", `Quick, link_ext_star;
-  ];
-  "Media Type", [
-    "Media Type", `Quick, get_media_type;
-  ];
-  "Auth", [
-    "Valid Auth", `Quick, valid_auth;
-  ];
-  "Cookie", [
-    "Valid Set-Cookie", `Quick, valid_set_cookie;
-    "Valid Cookie", `Quick, valid_cookie;
-    "Cookie with =", `Quick, cookie_with_eq_val;
-    "Ignores empty cookie", `Quick, ignores_empty_cookie;
-  ];
-  "Content Range", [
-    "none", `Quick, Content_range.none;
-    "content-length", `Quick, Content_range.content_length;
-    "content-range", `Quick, Content_range.content_range;
-  ];
-  "Cache Control", [
-    "concat", `Quick, test_cachecontrol_concat
-  ];
-  "Header", [
-    "get list valued", `Quick, list_valued_header;
-    "trim whitespace", `Quick, trim_ws;
-    "large header", `Slow, large_header;
-    "many headers", `Slow, many_headers;
-  ];
-]
+let transfer_encoding () =
+  let h = H.of_list ["transfer-encoding", "gzip"; 
+                     "transfer-encoding", "chunked"] in
+  let sh = H.to_string h in
+  aes "transfer_encoding_string_is_ordered"
+    sh "transfer-encoding: gzip\r\ntransfer-encoding: chunked\r\n\r\n";
+  let sh = H.get h "transfer-encoding" in
+  aeso "transfer_encoding_get_is_ordered" (Some "gzip,chunked") sh
+
+let () = Printexc.record_backtrace true
+
+let () =
+  Alcotest.run "test_header" [
+    "Link", [
+      "simple", `Quick, link_simple;
+      "multiple rels", `Quick, link_multi_rel;
+      "multiple lines", `Quick, link_multi_line;
+      "multiheader", `Quick, link_multi_multi;
+      "rel uri", `Quick, link_rel_uri;
+      "anchor", `Quick, link_anchor;
+      "rev", `Quick, link_rev;
+      "media", `Quick, link_media;
+      "media complex", `Quick, link_media_complex;
+      "title", `Quick, link_title;
+      "title star", `Quick, link_title_star;
+      "type token", `Quick, link_type_token;
+      "type quoted", `Quick, link_type_quoted;
+      "extension", `Quick, link_ext;
+      "extension star", `Quick, link_ext_star;
+    ];
+    "Media Type", [
+      "Media Type", `Quick, get_media_type;
+    ];
+    "Auth", [
+      "Valid Auth", `Quick, valid_auth;
+    ];
+    "Cookie", [
+      "Valid Set-Cookie", `Quick, valid_set_cookie;
+      "Valid Cookie", `Quick, valid_cookie;
+      "Cookie with =", `Quick, cookie_with_eq_val;
+      "Ignores empty cookie", `Quick, ignores_empty_cookie;
+    ];
+    "Content Range", [
+      "none", `Quick, Content_range.none;
+      "content-length", `Quick, Content_range.content_length;
+      "content-range", `Quick, Content_range.content_range;
+    ];
+    "Cache Control", [
+      "concat", `Quick, test_cachecontrol_concat
+    ];
+    "Header", [
+      "get list valued", `Quick, list_valued_header;
+      "trim whitespace", `Quick, trim_ws;
+      "replace existing", `Quick, Updates.replace_headers_if_exists;
+      "replace absent", `Quick, Updates.replace_headers_if_absent;
+      "update existing", `Quick, Updates.update_headers_if_exists;
+      "update existing list", `Quick, Updates.update_headers_if_exists_multi;
+      "update add absent", `Quick, Updates.update_headers_if_absent_add;
+      "update rm existing", `Quick, Updates.update_headers_if_exists_rm;
+      "update rm absent", `Quick, Updates.update_headers_if_absent_rm;
+      "update absent", `Quick, Updates.update_headers_if_absent;
+      "many headers", `Slow, many_headers;
+      "transfer encoding is in correct order", `Quick, transfer_encoding;
+    ]
+    @ if Sys.word_size = 64 then ["large header", `Slow, large_header] else []
+  ]