+++ cohttp/test/test_request.ml
@@ -45,6 +45,40 @@ let auth_uri _ =
     (r |> Request.headers |> Header.get_authorization)
     (Some (`Basic ("foo", "bar%25")))
 
+let t_encoding =
+  Alcotest.testable
+    (fun fmt e ->
+       let sexp = Cohttp.Transfer.sexp_of_encoding e in
+       Sexplib0.Sexp.pp fmt sexp) (=)
+
+let encoding_content_length_header () =
+  let r = Request.make ~headers:(Cohttp.Header.of_list ["content-length", "100"]) (Uri.of_string "http://someuri.com")  in
+  Alcotest.check t_encoding
+    "body encoding determined by content-length header"
+    (r |> Request.encoding)
+    (Fixed 100L)
+
+let encoding_transfer_encoding_header () =
+  let r = Request.make ~headers:(Cohttp.Header.of_list ["transfer-encoding", "chunked"]) (Uri.of_string "http://someuri.com")  in
+  Alcotest.check t_encoding
+    "body encoding determined by transfer-encoding header"
+    (r |> Request.encoding)
+    (Chunked)
+
+let encoding_both_headers () =
+  let r = Request.make ~headers:(Cohttp.Header.of_list ["transfer-encoding", "chunked"; "content-length", "100"]) (Uri.of_string "http://someuri.com")  in
+  Alcotest.check t_encoding
+    "body encoding with content-length and transfer-encoding headers."
+    (r |> Request.encoding)
+    (Chunked)
+
+let encoding_header_opt_argument () =
+  let r = Request.make ~encoding:Chunked ~headers:(Cohttp.Header.of_list ["content-length", "100"]) (Uri.of_string "http://someuri.com")  in
+  Alcotest.check t_encoding
+    "body encoding with content-length and transfer-encoding headers."
+    (r |> Request.encoding)
+    (Fixed 100L)
+
 let opt_default default = function
   | None -> default
   | Some v -> v
@@ -56,6 +90,10 @@ module Parse_result = struct
     | `Ok x -> `Ok (f x)
     | (`Invalid _ | `Eof) as e -> e
 end
+
+let uri_testable
+  : Uri.t Alcotest.testable
+  = Alcotest.testable Uri.pp_hum Uri.equal
 
 let t_parse_result_uri
   : Uri.t Parse_result.t Alcotest.testable
@@ -206,42 +244,55 @@ let parse_request_uri_host_traversal _ =
   let uri = `Ok (Uri.of_string "//example.com/etc/shadow") in
   parse_request_uri_ r uri "parse_request_uri_host_traversal"
 
-;;
-Printexc.record_backtrace true;
-Alcotest.run "test_request" [
-  "Auth", [
-    "header has auth", `Quick, header_has_auth;
-    "URI has user info", `Quick, uri_has_userinfo;
-    "from URI - do not override", `Quick, auth_uri_no_override;
-    "from URI", `Quick, auth_uri;
-  ];
-  "Parse URI", [
-    "simple", `Quick, parse_request_uri;
-    "with host", `Quick, parse_request_uri_host;
-    "with host and port", `Quick, parse_request_uri_host_port;
-    "double slash", `Quick, parse_request_uri_double_slash;
-    "double slash with host", `Quick, parse_request_uri_host_double_slash;
-    "triple slash", `Quick, parse_request_uri_triple_slash;
-    "triple slash with host", `Quick, parse_request_uri_host_triple_slash;
-    "no slash", `Quick, parse_request_uri_no_slash;
-    "no slash with host", `Quick, parse_request_uri_host_no_slash;
-    "empty", `Quick, parse_request_uri_empty;
-    "empty with host", `Quick, parse_request_uri_host_empty;
-    "path like scheme", `Quick, parse_request_uri_path_like_scheme;
-    "path like scheme with host", `Quick, parse_request_uri_host_path_like_scheme;
-    "path like host:port", `Quick, parse_request_uri_path_like_host_port;
-    "path like host:port with host",
-    `Quick, parse_request_uri_host_path_like_host_port;
-    "with query string", `Quick, parse_request_uri_query;
-    "with query with host", `Quick, parse_request_uri_host_query;
-    "no slash with query string", `Quick, parse_request_uri_query_no_slash;
-    "no slash with query with host",
-    `Quick, parse_request_uri_host_query_no_slash;
-    "CONNECT", `Quick, parse_request_connect;
-    "CONNECT with host", `Quick, parse_request_connect_host;
-    "OPTIONS", `Quick, parse_request_options;
-    "OPTIONS with host", `Quick, parse_request_options_host;
-    "parent traversal", `Quick, parse_request_uri_traversal;
-    "parent traversal with host", `Quick, parse_request_uri_host_traversal;
-  ];
-]
+let uri_round_trip _ =
+  let expected_uri = Uri.of_string "https://www.example.com/test" in
+  let actual_uri = Request.make expected_uri |> Request.uri in
+  Alcotest.check uri_testable "Request.make uri round-trip" actual_uri expected_uri
+
+let () = Printexc.record_backtrace true
+
+let () =
+  Alcotest.run "test_request" [
+    "Auth", [
+      "header has auth", `Quick, header_has_auth;
+      "URI has user info", `Quick, uri_has_userinfo;
+      "from URI - do not override", `Quick, auth_uri_no_override;
+      "from URI", `Quick, auth_uri;
+    ];
+    "Encoding", [
+      "from content-length header",`Quick, encoding_content_length_header;
+      "from transfer-encoding header",`Quick, encoding_transfer_encoding_header;
+      "with both headers",`Quick, encoding_both_headers;
+      "from both optional argument and headers", `Quick, encoding_header_opt_argument;
+    ];
+    "Parse URI", [
+      "simple", `Quick, parse_request_uri;
+      "with host", `Quick, parse_request_uri_host;
+      "with host and port", `Quick, parse_request_uri_host_port;
+      "double slash", `Quick, parse_request_uri_double_slash;
+      "double slash with host", `Quick, parse_request_uri_host_double_slash;
+      "triple slash", `Quick, parse_request_uri_triple_slash;
+      "triple slash with host", `Quick, parse_request_uri_host_triple_slash;
+      "no slash", `Quick, parse_request_uri_no_slash;
+      "no slash with host", `Quick, parse_request_uri_host_no_slash;
+      "empty", `Quick, parse_request_uri_empty;
+      "empty with host", `Quick, parse_request_uri_host_empty;
+      "path like scheme", `Quick, parse_request_uri_path_like_scheme;
+      "path like scheme with host", `Quick, parse_request_uri_host_path_like_scheme;
+      "path like host:port", `Quick, parse_request_uri_path_like_host_port;
+      "path like host:port with host",
+      `Quick, parse_request_uri_host_path_like_host_port;
+      "with query string", `Quick, parse_request_uri_query;
+      "with query with host", `Quick, parse_request_uri_host_query;
+      "no slash with query string", `Quick, parse_request_uri_query_no_slash;
+      "no slash with query with host",
+      `Quick, parse_request_uri_host_query_no_slash;
+      "CONNECT", `Quick, parse_request_connect;
+      "CONNECT with host", `Quick, parse_request_connect_host;
+      "OPTIONS", `Quick, parse_request_options;
+      "OPTIONS with host", `Quick, parse_request_options_host;
+      "parent traversal", `Quick, parse_request_uri_traversal;
+      "parent traversal with host", `Quick, parse_request_uri_host_traversal;
+      "uri round-trip", `Quick, uri_round_trip;
+    ];
+  ]